Convert Switch in Copy Service to TypeScript
